+++ versions/versions.h
@@ -22,7 +22,7 @@
 #ifndef _VERSION_H
 #define _VERSION_H
 
-#define CHUNK_VERSION "1.0.1"
+#define CHUNK_VERSION "1.1.0"
 #define CONCORDANCE_VERSION "1.1.0"
 #define LIGATE_VERSION "1.1.0"
 #define PHASE_VERSION "1.1.0"