+++ versions/CHANGELOG.md
@@ -15,6 +15,8 @@
 
 ### GLIMPSE chunk
 
+* v1.1.0
+	* Lifted version to v1.1.0
 * v1.0.1
 	* Removed option (--input) requiring input dataset and merged with --reference. Only reference panel is now required using the --input option.
 * v1.0.0
@@ -25,12 +27,15 @@
 * v1.1.0
 	* Performs indexing on input files if not indexed (avoiding to run bcftools after GLIMPSE_phase)
 	* Added additional checks for FPLOIDY
+	* Added haploid/diploid/mixed ploidy support
 
 * v1.0.0
 	* First release. Version used for the paper
 
 ### GLIMPSE sample
 
+* v1.1.0
+	* Added haploid/diploid/mixed ploidy support
 * v1.0.0
 	* First release. Version used for the paper